Pytorch-UNet-master/requirements.txt
matplotlib~=3.2.1
numpy~=1.18.2
Pillow~=7.1.2
torch~=1.4.0
torchvision~=0.5.0
tensorboard
future

opencv-python~=4.2.0.34
scikit-image~=0.17.2
tqdm~=4.46.0